% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm.R
\docType{data}
\name{atrisk}
\alias{atrisk}
\title{Data on TN endangered species}
\format{An object of class \code{list} of length 85.}
\usage{
atrisk
}
\description{
This is a list created with the lr_cache_all() function.
Each element of that list has two elements: species_name with the name of a federally-listed endangered or threatened species in Tennessee, and a prediction element that has the result of the lr_prediction function.
Note that the location data and prediction is for the species' entire range, not just the range in Tennessee.
Species with insufficient data for analysis are not included
}
\author{
Brian O'Meara \email{omeara.brian@gmail.com}
}
\keyword{data}
