% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{lr_get_gbif_data_for_species}
\alias{lr_get_gbif_data_for_species}
\title{Get GBIF data for a given species}
\usage{
lr_get_gbif_data_for_species(species, limit = 1000, country = "US",
  state = NULL)
}
\arguments{
\item{species}{The species name to search for info for}

\item{limit}{How many records to return}

\item{country}{Two letter country code. NULL if you want all countries.}

\item{state}{Full state name. NULL if you want all states.}
}
\value{
An object of class gbif (from the rgbif package)
}
\description{
Download data for a given species from GBIF
}
