% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{lr_get_listed_species}
\alias{lr_get_listed_species}
\title{Get endangered and threatened US species}
\usage{
lr_get_listed_species(state = "TN")
}
\arguments{
\item{state}{Two letter state code}
}
\value{
A vector of species
}
\description{
Downloads all species listed on https://ecos.fws.gov's site for a state
}
