% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm.R
\name{lr_prediction}
\alias{lr_prediction}
\title{Create species model from GBIF object}
\usage{
lr_prediction(gbif)
}
\arguments{
\item{gbif}{gbif class object, from lr_get_gbif_data_for_species() or rgbif::occ_search()}
}
\value{
list with objects prediction (from dismo::predict()), geographic_extent (raster extent return), observations (data.frame of longitudes and latitudes), ranges (vector of ranges of species data), and gbif (original data).
}
\description{
Uses ideas from https://jcoliver.github.io/learn-r/011-species-distribution-models.html, by Jeff Oliver
}
