% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm.R
\name{lr_prediction_plot}
\alias{lr_prediction_plot}
\title{Plot prediction and species points}
\usage{
lr_prediction_plot(lr_prediction_output, state = "tennessee",
  xlim = c(-90.360481, -81.53846), ylim = c(34.947001, 36.67528))
}
\arguments{
\item{lr_prediction_output}{Output from lr_prediction_output() function}

\item{state}{State to plot (for all states, pass in ".")}

\item{xlim}{Min and max of longitude}

\item{ylim}{Min and max of latitude}
}
\value{
Nothing
}
\description{
Follows Jeff Oliver's tutorial at https://jcoliver.github.io/learn-r/011-species-distribution-models.html
}
\examples{
prediction_object <- lr_prediction(lr_get_gbif_data_for_species("Quercus alba", country="US", state="Tennessee"))
lr_prediction_plot(prediction_object)
}
