% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm.R
\name{lr_cache_all}
\alias{lr_cache_all}
\title{Function to cache all species data}
\usage{
lr_cache_all(pkg = TRUE)
}
\arguments{
\item{pkg}{If TRUE, save within package}
}
\value{
A list of lists. Each element in the list has a species_name and the prediction and other information from lr_prediction()
}
\description{
Function to cache all species data
}
